Updates from Ruchith Fernando <ruchith.fernando@gmail.com>
These fixed the problems (xml sec init. stuff) with the two unit test cases.
  * Derived Key Encrypt/Decrypt
  * Derived Key Sign/Verify

And there are updates for the two Secure Conversation handlers which
I'm working on to get the existing sample working, so that the guys
working on the WS-RM/SC/T interop can use it as guide.