Added the MessageContext parameter to getClassLoader() method
Axis2 handlers will have to hold on to the RequestData  object as an attribute of the handler classes if they are to return the ServiceClass loader. This causes problems in multi threaded environments and simply breaks the stateless handler concept. Therefore we have to pass in the MessageContext object into the getClassloader() method to figure out the classloader.