Modifying the handlers to specify the crypto implementation dynamically
Right now WSS4J allows users to specify the implementation of 
org.apache.ws.security.components.crypto.Crypto to be specified using a property 
file. I'd like to propose a few changes to optionally load provide a reference 
to a java.util.Properties object with out specifying the .properties file.

This requires an extra method in the CryptoFactory class to create a Crypto 
instance using a Properties object.

public static Crypto getInstance(Properties properties) 

And then changes to the following methods in WSHandler to look for the Properties 
object in the MessageContext:

loadSignatureCrypto(RequestData)
loadEncryptionCrypto(RequestData)
loadDecryptionCrypto(RequestData)

In the case where a client wants to use a Properties object rather than a 
property file, he/she should create that object and add that to the
MessageContext under an id that he/she picks. This id MUST be specified in either
one of the handler parameters named (In axis2-rampart configuration these will 
be child elements within the <action> element):

SignaturePropRefId
encryptionPropRefId
decryptionPropRefId